Reyna að debugga divide by zero villu

+++ Models/ViewModels/BookView.cs
@@ -12,7 +12,7 @@ namespace BookCave.Models.ViewModels
         public int year { get; set; }
         public int numberOfPages { get; set; }
         public double rating { get; set; }
-        public int noOfRatings { get; set; }
+        public double noOfRatings { get; set; }
         public string description { get; set; }
         public string country { get; set; }
         public string language { get; set; }
